Make it easier to enable the Django Debug Toolbar, and make it work with CSP
